Added Authorization to protect reviews and campground routes

+++ routes/campgrounds.js
@@ -1,21 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const {campgroundSchema} = require('../schema');
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
 const Campground = require('../models/campground');
-const {isLoggedIn} = require('../utils/middleware');
+const {isLoggedIn, isAuthor, validateCampground} = require('../utils/middleware');
 
-const validateCampground = (req, res, next) => {
-   
-    const {error} = campgroundSchema.validate(req.body);
-    if(error){
-        const msg = error.details.map(ele => ele.message).join(',');
-        throw new ExpressError(msg, 400);
-    }
-    else
-        next();
-}
 
 router.get('/', catchAsync(async(req, res) => {
     const campgrounds = await Campground.find({});
@@ -28,7 +16,13 @@ router.get('/new',isLoggedIn, (req, res) => {
 
 router.get('/:id', catchAsync(async (req, res) =>{
     const {id} = req.params;
-    const campground = await Campground.findById(id).populate('reviews');
+    const campground = await Campground.findById(id).populate({
+        path : 'reviews',
+        populate : {
+            path : 'author'
+        }
+        })
+        .populate('author');
     if(!campground){
         req.flash('error', 'Cannot find the Campground');
         return res.redirect('/campgrounds');
@@ -36,7 +30,7 @@ router.get('/:id', catchAsync(async (req, res) =>{
     res.render('campgrounds/show', {campground});
 }))
 
-router.get('/:id/edit',isLoggedIn, catchAsync(async (req, res) =>{
+router.get('/:id/edit',isLoggedIn, isAuthor, catchAsync(async (req, res) =>{
     const {id} = req.params;
     const campground = await Campground.findById(id);
     if(!campground){
@@ -46,14 +40,15 @@ router.get('/:id/edit',isLoggedIn, catchAsync(async (req, res) =>{
     res.render('campgrounds/edit', {campground});
 }))
 
-router.post('/',validateCampground,isLoggedIn, catchAsync(async (req, res) => {
+router.post('/', validateCampground, isLoggedIn, catchAsync(async (req, res) => {
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash('success', 'Succesfully made a new campground');
     res.redirect(`campgrounds/${campground._id}`);
 }))
 
-router.put('/:id', isLoggedIn, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const {id} = req.params;
     const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground});
     req.flash('success', 'Successfully updated campground');
@@ -61,8 +56,13 @@ router.put('/:id', isLoggedIn, catchAsync(async (req, res) => {
     res.redirect(`${campground._id}`);
 }))
 
-router.delete('/:id', isLoggedIn, catchAsync(async(req, res) => {
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync(async(req, res) => {
     const {id} = req.params;
+    const campground = await Campground.findById(id);
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error', 'You are not authorised to update this campground');
+        return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success','Successfully deleted campground');
     res.redirect("/campgrounds");
